add new nesc extreme wind load case [finishes #185711253]
