+++ src/main/groovy/com/spidasoftware/schema/conversion/changeset/ConverterUtils.groovy
@@ -69,8 +69,8 @@ class ConverterUtils {
         converter.addChangeSet(9, new ClearancesChangeset())
         converter.addChangeSet(10, new IceDensityChangeSet())
         converter.addChangeSet(10, new TerrainLayerChangeSet())
-        converter.addChangeSet(10, new CSAMaxWindLoadCaseChangeSet())
         converter.addChangeSet(10, new LoadCaseChangeSet())
+        converter.addChangeSet(10, new LoadCaseNameChangeSet())
 
         // add calc changesets above here
         converter.setCurrentVersion(currentVersion)
@@ -85,8 +85,8 @@ class ConverterUtils {
         converter.addChangeSet(9, new ExtremeWindLoadCaseChangeset())
         converter.addChangeSet(9, new ClearancesChangeset())
         converter.addChangeSet(10, new IceDensityChangeSet())
-        converter.addChangeSet(10, new CSAMaxWindLoadCaseChangeSet())
         converter.addChangeSet(10, new LoadCaseChangeSet())
+        converter.addChangeSet(10, new LoadCaseNameChangeSet())
         // add client data changesets above here
 
         converter.setCurrentVersion(currentVersion)
@@ -101,8 +101,8 @@ class ConverterUtils {
         converter.addChangeSet(9, new ExtremeWindLoadCaseChangeset())
         converter.addChangeSet(9, new ClearancesChangeset())
         converter.addChangeSet(10, new IceDensityChangeSet())
-        converter.addChangeSet(10, new CSAMaxWindLoadCaseChangeSet())
         converter.addChangeSet(10, new LoadCaseChangeSet())
+        converter.addChangeSet(10, new LoadCaseNameChangeSet())
         // add result changesets above here
 
         converter.setCurrentVersion(currentVersion)
